changed CodeSamples to TestMethodBuilder
