+++ ataru.gemspec
@@ -19,9 +19,9 @@ Gem::Specification.new do |spec|
   spec.require_paths = ["lib"]
 
   spec.add_development_dependency "bundler", "~> 1.5"
-  spec.add_development_dependency "rake"
+  spec.add_development_dependency "rake", "~> 10.3"
 
-  spec.add_runtime_dependency "minitest"
-  spec.add_runtime_dependency "kramdown"
-  spec.add_runtime_dependency "thor"
+  spec.add_runtime_dependency "minitest", "~> 5.4"
+  spec.add_runtime_dependency "kramdown", "~> 1.4"
+  spec.add_runtime_dependency "thor", "~> 0.19"
 end
