+++ lib/ataru.rb
@@ -1,8 +1,9 @@
 require 'ataru/version'
 require 'ataru/markdown_loader'
 require 'ataru/traverser'
-require 'ataru/code_samples'
+require 'ataru/code_sample'
 require 'ataru/argument_checker'
 require 'ataru/test_converter'
+require 'ataru/test_method_builder'
 require 'ataru/application'
 require 'ataru/cli/application'
